chore: same thing with deloy logic
